Adding parser functionality
There are two large changes.

1) python parser front end and code generation is added to pdaggerq. This
allows a user to generate CC codes.  Worked examples for code generation modules
interfacing with the C++ pdaggerq module are provided in examples.  In
examples/full_cc_codes we provide a pedagogical example of CCSD and lambda-CCSD
codes.

2) The module generated by pybind is now called _pdaggerq.  The reason for the
change is that inorder to have a python module named pdaggerq that contained
the C++ module code and the python parsing code we need to change the name
of the C++ module.  Following python conventions we made the module generated
by pybind _pdaggerq and import ._pdaggerq * in the __init__.py of the
pdaggerq module containing the python code.