Add method to check if vertex is temporary and improve logic.
Added method `is_temp` to check if a vertex is a temporary object. Made improvements to graph making group together nodes that are contracted and their external lines. The `Term::is_compatible` function and the function in the substitute functions incorrectly filter out invalid contractions. these now always return true, but the logic needs to be fixed to get a better speedups.