Refactor and clean-up of Linkage class functionality.
Removed code relating building an iterator for the linkage class that was a bust. This should be implemented in principle, but for now removing it reduces clutter.

Refactored the Term class to test the new scaling of a term without affecting the original term. This prevents having to construct a new Term object to find the modified scaling. Made further changes to representation of the graph structure of equations in dot_generator.cc. Other changes are temporary fixes while logic is being debugged.