Improved ordering of intermediates such that nested intermediates show up immediately after each other. This prevents storing the original intermediate in memory for too long. Also added logic to remove intermediates that only show up once after it is nested into another intermediate.
Cleaned up some of the options and their documentation. Added the ability to specify the max size of an intermediate to prevent storage of too large containers if the user desires.

Fixed functionality of `Linkage::to_vector` when requesting to not fully expand the linkage chain. To do this, I added another mutable container for lazy evaluation of the nonfully expanded term.

Refactored the Line class to have more descriptive names for the occupied/virtual status and the block type

Fixed some issues with rank 1 containers for sigma build being considered as a scalar. Also added functionality to remove the sigma index during optimization. Next commit will separate tmps with a sigma index separately.