Fixed issue with sorting and printing of blocked equation (either by spin or range).
This fix required modifications to the recursive expansion of the linkage.cc class. The most notable change was that the declaration of the righthand side tensors of the Term instances created from a LinkagePtr would set the children to be the `left` and `right` of the linkage. This worked since the operators would be expanded ultimately, but ideally the rhs should be generated recursively, which is now the case.