Refactor functions and improve permutation clarity
term_perm_helper.cc was significantly refactored for readability and now handles permuting terms with linkages consistently.

The `to_vector` functions was removed entirely since the recursive process of getting the vertices was a bottleneck. the list of vertices are now built during the construction of the linkage and are saved as a data-member. the new function name to get these vertices is `get_vertices`.