Refactor code for more efficient vertex and line comparisons
The linkage vertices are now efficiently stored as a vector instead of an unordered set, improving the search and comparison operations. The scaling of operations is now also handled from a central location, eliminating the need to calculate it multiple times. Hashing now includes the consideration of label permutations. Removed redundant tracking of variables which were replaced with on-demand calculations.