substituted vertices in term_substitute.cc are now re-ordered instead of re-computing the scaling.
now using a multimap instead of a vector of pairs relating flop scaling maps to linkages. Connections between lines in linkage.h are now stored in a vector of arrays instead of a vector of pairs for better performance. Instead of inserting the connections in order, we change the equality operator to check if all pairs from one linkage exists in the other.