Changed functionality of find_links to find all occurrences of the link.
refactored `PQGraph::make_scalars()` into the `substitute.cc` file.

Added `PQGraph::reindex` to reindex all the intermediates in order of occurrence

Changed substitution algorithm to include additions

changed default in `Equation::rearrange(type)` to 'a' so that it rearranges the equations first by the scalars, then the reused intermediates, and finally by the non-reusable intermediates.

Changed `linkage_set::insert(...)` to take in a VertexPtr instead of a LinkagePtr for the sake of generality. TODO: refactor the linkage_set to also hash pure vertices.