`Linkage::fuse()` would not correctly generate fusions for all test cases. The issue was the the order of the lines would change which would cause problems for nested contractions. a check for the order of the lines has been added and different ordering of the fusion operations are taken until lines match (if at all)
Also replaced shallow copies of Linkages with a return of the linkage itself wherever possible.

also required was to not sort the lines of the linkages upon construction, since the order should not change when applying fusions. perhaps this could be remedied by sorting the lines during the printing phase.